Fix type warnings
typescript in management app was throwing:
Property 'getData' does not exist on type 'DirectoryEntry'.

the way types were setup in zip.js, the entry is a union type and only
typed as a DirectoryEntry if it had a .directory property, and as
FileEntry when .file was present.  Checking the properties allows
typescript to infer types as it should